+++ api/views.py
@@ -1,15 +1,13 @@
 from django.core.validators import URLValidator
 from django.core.exceptions import ValidationError
 from django.http import HttpResponse
-from rest_framework import viewsets, generics, exceptions
-from rest_framework.permissions import IsAuthenticated
+from rest_framework import viewsets, generics
 from rest_framework.response import Response
 from rest_framework.decorators import list_route
-from knox.auth import TokenAuthentication
 from shortener.models import shorten
-from api.models import Member, Team, Role, Update, Event
+from accounts.auth import LabsAuthMixin
+from api.models import Member, Team, Role, Update
 from api.serializers import MemberSerializer, TeamSerializer, RoleSerializer, UpdateSerializer
-from api.auth import LabsTokenAuthentication
 
 
 class ShortURLViewSet(generics.GenericAPIView):
@@ -57,16 +55,6 @@ class UpdateViewSet(viewsets.ModelViewSet):
     queryset = Update.objects.all()
     serializer_class = UpdateSerializer
     http_method_names = ['get']
-
-
-class PennAuthMixin(generics.GenericAPIView):
-    authentication_classes = (TokenAuthentication,)
-    permission_classes = (IsAuthenticated,)
-
-
-class LabsAuthMixin(generics.GenericAPIView):
-    authentication_classes = (LabsTokenAuthentication,)
-    permission_classes = (IsAuthenticated,)
 
 
 class ProtectedViewSet(LabsAuthMixin, generics.GenericAPIView):
